+++ src/main/java/com/example/springboot/student/StudentController.java
@@ -26,13 +26,13 @@ public class StudentController {
   }
 
   @DeleteMapping(path = "{studentId}")
-  public void deleteStudent(@PathVariable("studentId") Long studentId) {
+  public void deleteStudent(@PathVariable("studentId") Integer studentId) {
     studentService.deleteStudent(studentId);
   }
 
   @PutMapping(path = {"{studentId}"})
   public void updateStudent(
-      @PathVariable("studentId") Long studentId,
+      @PathVariable("studentId") Integer studentId,
       @RequestParam(required = false) String name,
       @RequestParam(required = false) String email) {
     studentService.updateStudent(studentId, name, email);
